move source code to src
