+++ docs/conf.py
@@ -13,7 +13,7 @@
 import sys
 from pathlib import Path
 
-sys.path.insert(0, str(Path("..").resolve()))
+sys.path.insert(0, str(Path("../src").resolve()))
 
 
 # -- Project information -----------------------------------------------------
