--- pystellibs/interpolator/README.rst
@@ -1,3 +1,0 @@
-Interpolator for the different spectral libraries.
-
-Some interpolation are very generic and can be used for multiple libraries.
